For #16477: Switch to using an extension to update a11y collection info.
